add encoder decoder example
